Switch to using 64 bit floats by default.

+++ pylag/data_types_python.py
@@ -1,4 +1,4 @@
 import numpy as np
 
-DTYPE_FLOAT = np.float32
-DTYPE_INT = np.int32
+DTYPE_FLOAT = np.float64
+DTYPE_INT = np.int64
